requirements.txt
# ============================================
# EcoSort IA — Dependencias
# ============================================

# --- Entrenamiento (correr en laptop/PC/Colab) ---
tensorflow>=2.15.0
keras>=3.0.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# --- Inferencia en Raspberry Pi 5 ---
# En la RPi instalar solo:
#   pip install tflite-runtime opencv-python-headless numpy RPi.GPIO
# NO instalar tensorflow completo en la RPi (es muy pesado)

# --- Dashboard ---
flask>=3.0.0

# --- Utilidades ---
opencv-python>=4.8.0
